[EL-84] PUSH-уведомления. Часть 1

+++ feature/notifications/implementation/src/main/java/ru/dekabrsky/feature/notifications/implementation/data/model/NotificationDbEntity.kt
@@ -1,16 +1,12 @@
 package ru.dekabrsky.feature.notifications.implementation.data.model
 
-import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import androidx.room.TypeConverters
-import ru.dekabrsky.feature.notifications.implementation.data.converters.DateConverter
-import java.util.Date
 
 @Suppress("LongParameterList") // todo я конечно дико извиняюсь, но за что?
 @Entity(tableName = NotificationDbEntity.TABLE_NOTIFICATIONS)
 class NotificationDbEntity(
-    @PrimaryKey val uid: Int? = null,
+    @PrimaryKey val uid: Long? = null,
     val tabletName: String,
     val dosage: String?,
     val note: String?,
